+++ griptape/drivers/vector/astradb_vector_store_driver.py
